Finished read_int(), working on rest of built in functions.

+++ interpreter/headers/value.h
@@ -5,7 +5,7 @@ typedef struct VALUE
 {
   int type;
   union {
-    int integer;
+    long int integer;
     int boolean;
     char* string;
   } v;
